feat: implement assignments  modular components and management modals
